fix: Split input validation into another file
